requirements.txt
fastapi~=0.115.8
uvicorn~=0.34.0
pymongo
pymongo[srv]
python-dotenv
pydantic~=2.10.6
motor~=3.7.0
opencv-python~=4.11.0.86
numpy~=2.0.2
Flask~=3.1.0
websockets~=14.2
scipy~=1.15.1
torch~=2.6.0+cu126
pillow~=11.1.0
transformers~=4.48.2
imutils~=0.5.4
dlib~=19.24.6